Move ChildShape to separate file
